Make fluxcloud port configurable to fix #16
By default, we will still use port 3031 to prevent breaking users that update fluxcloud
with Flux, since their services will all still point at 3031 we can't change the default
in the image.

We make the port configurable by the LISTEN_ADDRESS configuration option and set all
default configurations to 3032 to avoid a conflict when configuring Flux with the
recommended metrics port of 3031.